add an example of single subject rois; reduce example time by allowing them to share derivatives not dependent on tractography, such as mapping, csd, etc
