+++ AFQ/data/s3bids.py
@@ -11,7 +11,6 @@ from dask import compute, delayed
 from dask.diagnostics import ProgressBar
 
 from pathlib import Path
-import os
 import os.path as op
 
 import logging
@@ -1220,44 +1219,6 @@ def s3fs_nifti_read(fname, fs=None, anon=False):
         fh = nib.FileHolder(fileobj=bb)
         img = nib.Nifti1Image.from_file_map({'header': fh, 'image': fh})
     return img
-
-
-def write_json(fname, data):
-    """
-    Write data to JSON file.
-
-    Parameters
-    ----------
-    fname : str
-        Full path to the file to write.
-
-    data : dict
-        A dict containing the data to write.
-
-    Returns
-    -------
-    None
-    """
-    with open(fname, 'w') as ff:
-        json.dump(data, ff, default=lambda obj: "Not Serializable")
-
-
-def read_json(fname):
-    """
-    Read data from a JSON file.
-
-    Parameters
-    ----------
-    fname : str
-        Full path to the data-containing file
-
-    Returns
-    -------
-    dict
-    """
-    with open(fname, 'r') as ff:
-        out = json.load(ff)
-    return out
 
 
 def s3fs_json_read(fname, fs=None, anon=False):
