+++ AFQ/definitions/mapping.py
@@ -7,7 +7,7 @@ import os.path as op
 from AFQ.definitions.utils import Definition, find_file
 from dipy.align import syn_registration, affine_registration
 import AFQ.registration as reg
-import AFQ.data.s3bids as afs
+from AFQ.utils.path import write_json
 from AFQ.tasks.utils import get_fname
 
 from dipy.align.imaffine import AffineMap
@@ -230,7 +230,7 @@ class GeneratedMapMixin(object):
             np.save(prealign_file, aff)
             meta_fname = get_fname(
                 base_fname, f'{prealign_file_desc}.json')
-            afs.write_json(meta_fname, meta)
+            write_json(meta_fname, meta)
         return prealign_file if save else np.load(prealign_file)
 
     def get_for_subses(self, base_fname, dwi, bids_info, reg_subject,
@@ -264,7 +264,7 @@ class GeneratedMapMixin(object):
                 meta["reg_subject"] = reg_subject
             if isinstance(reg_template, str):
                 meta["reg_template"] = reg_template
-            afs.write_json(meta_fname, meta)
+            write_json(meta_fname, meta)
         reg_prealign_inv = np.linalg.inv(reg_prealign) if self.use_prealign\
             else None
         mapping = reg.read_mapping(
