+++ examples/howto_examples/vof_example.py
@@ -16,7 +16,7 @@ import AFQ.utils.streamlines as aus
 
 import os.path as op
 
-afd.organize_stanford_data(clear_previous_afq=True)
+afd.organize_stanford_data(clear_previous_afq="track")
 
 bundle_names = ["ARC_L", "ARC_R", "pARC_L", "pARC_R", "VOF_L", "VOF_R"]
 bundle_dict = abd.BundleDict(bundle_names)
