+++ examples/tutorial_examples/plot_afq_api.py
@@ -41,10 +41,10 @@ import AFQ.data.fetch as afd
 # intializing the GroupAFQ object (which we will do next)
 #
 # The clear_previous_afq is used to remove any previous runs of the afq object
-# stored in the AFQ_data/stanford_hardi/ BIDS directory. Set it to false if
+# stored in the AFQ_data/stanford_hardi/ BIDS directory. Set it to None if
 # you want to use the results of previous runs.
 
-afd.organize_stanford_data(clear_previous_afq=True)
+afd.organize_stanford_data(clear_previous_afq="track")
 
 ##########################################################################
 # Set tractography parameters (optional)
